add skill prerequisite levels near the skill
