+++ my-app/src/components/skillsprite/SkillSprite.css
