+++ my-app/src/components/skillsummary/SkillSummary.css
@@ -1,0 +1,32 @@
+.SkillSummary {
+  display: flex;
+  flex-direction: column;
+}
+
+.SkillSummary-miniIcon {
+  transform: scale(0.25);
+  transform-origin: left top;
+  max-height: 25px;
+  aspect-ratio: 1;
+}
+
+.SkillSummary-levelValue {
+  margin-left: 4px;
+}
+
+.SkillSummary-itemLine {
+  display: flex;
+  align-items: left;
+  justify-content: left;
+  margin-left: 10%;
+}
+
+.SkillSummary-controls {
+
+}
+
+.SkillSummary-list {
+  display: flex;
+  flex-direction: column;
+  transition: height 0.5s;
+}
