+++ my-app/src/resources/skills.json
@@ -115,7 +115,7 @@
     "max": 10,
     "nextId": 1014,
     "nextLevel": 5,
-    "prevId": 1011,
+    "prevId": 1008,
     "prevLevel": 5,
     "spriteIndex": 11
   },
